requirements.txt
Django>=3.2
djangorestframework>=3.12.4
django-cors-headers==3.8.4
psycopg2==2.9.1
twilio==6.63.1
python-decouple==3.4